delete list bugfix
Changes made:
- Relocated `ListLabel` into components folder inside `src`.
- Fixed a situation that made the backend crash when deleting a list.
- Added sidebar scrolling for overflow.
- User now doesn't need to refresh page to see new lists list.
